+++ __test__/fixtures/enum-member-string-value-changed/new/index.ts
@@ -1,0 +1,4 @@
+export enum Status {
+  Active = 'pending',
+  Inactive = 'inactive',
+}