+++ __test__/fixtures/namespace-member-removed/new/index.ts
@@ -1,0 +1,5 @@
+export namespace Foo {
+  export interface Bar {
+    a(): void;
+  }
+}