+++ __test__/fixtures/namespace-member-removed/old/index.ts
@@ -1,0 +1,6 @@
+export namespace Foo {
+  export interface Bar {
+    a(): void;
+    b(): void;
+  }
+}