+++ __test__/fixtures/namespace-merged-function/new/index.ts
@@ -1,0 +1,7 @@
+export function F(): void {}
+
+export namespace F {
+  export interface Options {
+    a(): void;
+  }
+}