+++ __test__/fixtures/namespace-merged-function/old/index.ts
@@ -1,0 +1,8 @@
+export function F(): void {}
+
+export namespace F {
+  export interface Options {
+    a(): void;
+    b(): void;
+  }
+}