+++ __test__/fixtures/param-type-equivalent/new/index.ts
@@ -1,0 +1,1 @@
+export function foo(a: ReadonlyArray<string>): void {}