+++ __test__/fixtures/return-type-any-to-concrete/old/index.ts
@@ -1,0 +1,3 @@
+export function foo(): any {
+  return 1;
+}