+++ __test__/fixtures/variable-type-narrowed/old/index.ts
@@ -1,0 +1,1 @@
+export const config: string | number = 'a';